Rename Number analyzer to Number_analyzer.cpp
The program will gently handle faulty input and will not progress until the user gives a valid integer. In addition, the loop for calculating square roots has been optimized by computing the highest square root (integer part of the square root) of the user input and iterating only up to that value. This reduces the number of calculations and makes the software more efficient.